+++ CHANGELOG.md
@@ -12,6 +12,7 @@ Not yet released.
 * [CS-458][] `GET /commands/available` includes command and wrapper labels
 * [CS-29][] Enable output handlers to have type "`Assessor`". If a command output points to an assessor XML, the output handler can now upload that XML and create the assessor object in XNAT.
 * Manually refresh resource catalog when uploading resources. (We used to rely on the Catalog Service to do this, but it doesn't anymore.)
+* [CS-407][] Add docker host setting to translate mount paths. If xnat sees a path at /data/xnat/x/y/z but your docker host sees the same path at /my/path/x/y/z, you can include these path prefixes in your docker server settings. If you're using REST, set the properties "path-translation-xnat-prefix" and "path-translation-docker-prefix" respectively. If you're using the Plugin Settings UI, the fields will be there.
 
 ### Bugfixes
 
@@ -24,6 +25,7 @@ Not yet released.
 
 [CS-29]: https://issues.xnat.org/browse/CS-29
 [CS-80]: https://issues.xnat.org/browse/CS-80
+[CS-407]: https://issues.xnat.org/browse/CS-407
 [CS-457]: https://issues.xnat.org/browse/CS-457
 [CS-458]: https://issues.xnat.org/browse/CS-458
 [CS-480]: https://issues.xnat.org/browse/CS-480